refactor element details component
