+++ redisinsight/ui/src/pages/browser/modules/key-details/components/vector-set-details/element-details/ElementDetails.types.ts
@@ -1,0 +1,7 @@
+import { VectorSetElement } from 'uiSrc/slices/interfaces'
+
+export interface ElementDetailsProps {
+  element: VectorSetElement | null
+  isOpen: boolean
+  onClose: () => void
+}
