+++ redisinsight/ui/src/pages/browser/modules/key-details/components/vector-set-details/element-details/constants.ts
@@ -1,0 +1,4 @@
+export const VECTOR_DESCRIPTION =
+  'The numerical embedding representing this item in vector space, used for similarity search and ranking.'
+export const ATTRIBUTES_DESCRIPTION =
+  'Structured metadata associated with this item, used for filtering, display, and hybrid search queries.'
