+++ redisinsight/ui/src/pages/browser/modules/key-details/components/vector-set-details/element-details/utils.ts
@@ -1,0 +1,4 @@
+export const formatVector = (vector?: number[]): string => {
+  if (!vector?.length) return '[]'
+  return `[${vector.join(', ')}]`
+}
